+++ llm/test/README.md
@@ -1,0 +1,9 @@
+## 客户端
+ 
+- 支持open ai 两种api ：ChatCompletion 与 Completion
+
+### 使用方式
+- 这里提供了fastdelopy 客户端的示例demo，参考test 目录下 test_client.py
+- openai 两个接口示例代码在test 目录下 test_openai.py
+
+
